+++ frontend/src/components/Reusables/FooterComponents/FooterLinks.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import FooterData from "./FooterData"
 import twitter from '../../../assets/images/Twitter.png';
 import faceb from '../../../assets/images/Facebook.png';
