+++ frontend/src/pages/AboutUs/sections/FixIt.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import {
 	StyledFixHeader,
 	StyledFixImg,
