+++ frontend/src/pages/AboutUs/sections/Help.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import {
 	Blue,
 	Btn,
