+++ frontend/src/pages/AboutUs/sections/Hero.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import {
 	StyledHero,
 	StyledHeroMsg,
