+++ frontend/src/pages/AboutUs/sections/Story.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import {
 	StoryImg,
 	StyledStory,
