+++ frontend/src/pages/GlassDoor/GlassDoor.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import PageLayout from '../../layout/PageLayout';
 import {
 	HowWeWork,
